LayoutTests/imported/w3c/web-platform-tests/html/user-activation/activation-trigger-mouse-right.html

<!DOCTYPE html>
<html>
  <meta charset="utf-8" />
  <title>Test for right-click activation trigger.</title>
  <head>
    <link
      rel="help"
      href="https://html.spec.whatwg.org/multipage/interaction.html#activation-triggering-input-event"
    />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-actions.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <script src="resources/utils.js"></script>
  </head>
  <body>
    <h1>Just a h1</h1>
  </body>
  <script>
    promise_test(async () => {
      const actions = new test_driver.Actions();
      const mousedown_event = getEvent("mousedown");
      const mouseup_event = getEvent("mouseup");
      const auxclick_event = getEvent("auxclick");
      const contextmenu_event = getEvent("contextmenu");

      // In most non-Windows platforms the right-click context-menu appears on mousedown, so
      // mouseup and auxclick events are not received by the page if the menu is modal.  We
      // are suppressing the context-menu to guarantee receiving those later events.
      document.body.addEventListener("contextmenu", (e) => e.preventDefault());

      // Press down right mouse button.
      actions
        .pointerMove(0, 0, { origin: document.body })
        .pointerDown({ button: actions.ButtonType.RIGHT })
        .send();

      await mousedown_event;
      assert_true(
        navigator.userActivation.isActive,
        "mousedown event should result in activation"
      );
      await consumeTransientActivation();

      await contextmenu_event;
      assert_false(
        navigator.userActivation.isActive,
        "contextmenu should have no activation after mousedown consumption"
      );

      // Release right mouse button.
      actions.pointerUp({ button: actions.ButtonType.RIGHT }).send();

      await mouseup_event;
      assert_false(
        navigator.userActivation.isActive,
        "mouseup should have no activation after mousedown consumption"
      );

      await auxclick_event;
      assert_false(
        navigator.userActivation.isActive,
        "auxclick should have no activation after mousedown consumption"
      );
    }, "Activation through right-click mouse event");
  </script>
</html>
